Drop use of RIOT_EXPANDED_HEADERS
While the motivation of RIOT_EXPANDED_HEADERS was not as good as it was
before it was found that RIOT_CFLAGS had to be used to transport
compiler flags that influence layout, what killed it was the requirement
to announce the path to the riot-all.h file to RIOT's Makefile -- and
riot-all.h could live anywhere depending on where Cargo fetches the
crate from.

This now does all the expansion in Cargo, based on the flags that are
already present.

The version of bindgen was bumped as the 0.40.0 version caused troubles
with CFLAGS unknown to the compiler; presumably, the new version is
smarter there.