cstr: Pick an error type
The very generic impl previously used there for compatibility with a
future cstr_core pub exporting might have had a manual drop
implementation, and was thus usuitable for the use of cstr! inside const
expressions.

Adding an own FromBytesWithNulError that barely has any visible
properties allows use in const expressions, and can be replaced with a
`pub use` later. (Yes, there *may* be conflicting impls, but this is all
not as widespread as the Never type yet to expect practical trouble).