Work around missing intrinsics
The change of f4e73de2 only worked until --fail-on-error was enabled.
(I don't want to imagine which other kinds of previously silent errors
loom). Now that the error would be fatal, the offending functions are
preprocessor-renamed and redirected right in C (which helps c2rust get
the types right everywhere they're used).