Limit C2Rust in main module to explicitly listed items
This allows dropping workarounds necessary because items from
bindgen and C2Rust can interfere when in the same module (eg. by const
items from one part becoming patterns for the other).

This is a breaking change, as not all C2Rust items (in the inline
module) are exported any more. However, those commonly used (especially
by riot-wrappers) are on the explicit import list, so most applications
should work with the updated version as well.