build: Shparper matching for toplevel inclusion of C2Rust items
Previously, the `irq_disable()` function that's non-inline on native was
used in C2Rust transpiled code and thus referenced in an `extern "C" {}`
section, making the toplevel inline mechanism needlessly shadow the
bindgen provided function with an equally external C2Rust function.

With the more recent C2Rust versions' elision of `#[no_mangle]`, that
mechanism broke -- but it only worked in the first place because a relic
of earlier experimentation (towards using C2Rust exclusively) had made
the extern declaration static. Removing that line as it has long been
unused.