build: Tolerate absence of RIOT_USEMODULES
This is not a breaking change, it just turns what used to be a build.rs
error into a later error (eg. when used with an older riot-wrappers,
that would be missing modules -- but only if the build system suddenly
stopped passing in the variable).

Tolerating the absence allows the build system to phase out passing
dummy values into RIOT_USEMODULES even before [38] is through.

[38]: https://github.com/RIOT-OS/rust-riot-sys/pull/38